+++ packages/contracts/src/utils/ledger-utils.ts
@@ -14,15 +14,11 @@
  */
 
 import { getNetworkId } from '@midnight-ntwrk/midnight-js-network-id';
-import { type CompiledContract, ContractExecutable } from '@midnight-ntwrk/midnight-js-protocol/compact-js';
-import { type Contract, ProvableCircuitId, VerifierKey as ContractVerifierKey } from '@midnight-ntwrk/midnight-js-protocol/compact-js/effect/Contract';
 import {
   type AlignedValue,
-  type CoinPublicKey,
   type ContractAddress,
   ContractState,
   type QueryContext,
-  type SigningKey,
   type ZswapLocalState} from '@midnight-ntwrk/midnight-js-protocol/compact-runtime';
 import {
   ChargedState,
@@ -32,7 +28,6 @@ import {
   ContractState as LedgerContractState,
   type EncPublicKey,
   Intent,
-  type MaintenanceUpdate,
   type PartitionedTranscript,
   QueryContext as LedgerQueryContext,
   StateValue as LedgerStateValue,
@@ -44,12 +39,7 @@ import {
 } from '@midnight-ntwrk/midnight-js-protocol/ledger';
 import {
   type AnyProvableCircuitId,
-  asContractAddress,
-  asEffectOption,
-  makeContractExecutableRuntime,
-  Transaction,
-  type VerifierKey,
-  type ZKConfigProvider
+  Transaction
 } from '@midnight-ntwrk/midnight-js-types';
 import { assertDefined, ttlOneHour } from '@midnight-ntwrk/midnight-js-utils';
 
@@ -160,96 +150,3 @@ export const createUnprovenLedgerCallTx = (
     intent
   );
 };
-
-// Utilities for unproven transactions for the single contract updates above.
-
-export const unprovenTxFromContractUpdates = async (
-  updateAndSignFn: () => Promise<MaintenanceUpdate>
-): Promise<UnprovenTransaction> => {
-  return Transaction.fromParts(
-    getNetworkId(),
-    undefined,
-    undefined,
-    Intent.new(ttlOneHour()).addMaintenanceUpdate(await updateAndSignFn())
-  );
-};
-
-export const createUnprovenReplaceAuthorityTx = <C extends Contract.Any>(
-  zkConfigProvider: ZKConfigProvider<string>,
-  compiledContract: CompiledContract.CompiledContract<C, any>, // eslint-disable-line @typescript-eslint/no-explicit-any
-  contractAddress: ContractAddress,
-  newAuthority: SigningKey,
-  contractState: ContractState,
-  currentAuthority: SigningKey,
-  coinPublicKey: CoinPublicKey,
-): Promise<UnprovenTransaction> => {
-  const contractExec = ContractExecutable.make(compiledContract);
-  const contractRuntime = makeContractExecutableRuntime(zkConfigProvider, {
-    coinPublicKey,
-    signingKey: currentAuthority
-  });
-
-  return unprovenTxFromContractUpdates(async () => {
-    return (await contractRuntime.runPromise(contractExec.replaceContractMaintenanceAuthority(
-      asEffectOption(newAuthority),
-      {
-        address: asContractAddress(contractAddress),
-        contractState
-      }
-    ))).public.maintenanceUpdate
-  });
-}
-
-export const createUnprovenRemoveVerifierKeyTx = <C extends Contract.Any>(
-  zkConfigProvider: ZKConfigProvider<string>,
-  compiledContract: CompiledContract.CompiledContract<C, any>, // eslint-disable-line @typescript-eslint/no-explicit-any
-  contractAddress: ContractAddress,
-  operation: string,
-  contractState: ContractState,
-  currentAuthority: SigningKey,
-  coinPublicKey: CoinPublicKey,
-): Promise<UnprovenTransaction> => {
-  const contractExec = ContractExecutable.make(compiledContract);
-  const contractRuntime = makeContractExecutableRuntime(zkConfigProvider, {
-    coinPublicKey,
-    signingKey: currentAuthority
-  });
-
-  return unprovenTxFromContractUpdates(async () => {
-    return (await contractRuntime.runPromise(contractExec.removeContractOperation(
-      ProvableCircuitId(operation),
-      {
-        address: asContractAddress(contractAddress),
-        contractState
-      }
-    ))).public.maintenanceUpdate
-  });
-}
-
-export const createUnprovenInsertVerifierKeyTx = <C extends Contract.Any>(
-  zkConfigProvider: ZKConfigProvider<string>,
-  compiledContract: CompiledContract.CompiledContract<C, any>, // eslint-disable-line @typescript-eslint/no-explicit-any
-  contractAddress: ContractAddress,
-  operation: string,
-  newVk: VerifierKey,
-  contractState: ContractState,
-  currentAuthority: SigningKey,
-  coinPublicKey: CoinPublicKey,
-): Promise<UnprovenTransaction> => {
-  const contractExec = ContractExecutable.make(compiledContract);
-  const contractRuntime = makeContractExecutableRuntime(zkConfigProvider, {
-    coinPublicKey,
-    signingKey: currentAuthority
-  });
-
-  return unprovenTxFromContractUpdates(async () => {
-    return (await contractRuntime.runPromise(contractExec.addOrReplaceContractOperation(
-      ProvableCircuitId(operation),
-      ContractVerifierKey(newVk),
-      {
-        address: asContractAddress(contractAddress),
-        contractState
-      }
-    ))).public.maintenanceUpdate
-  });
-}
